moved stuff for editor team to a module so they can import it
